# openapi.gpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Captain Data MCP API",
    "description": "MCP API for managing Captain Data actions.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {},
...
  },
  "paths": {
...
    "/auth": {
      "post": {
        "operationId": "authenticate",
        "summary": "Authenticate with your Captain Data API key",
        "tags": [
          "Authentication"
        ],
        "description": "Exchange your Captain Data API key for a session token that can be used for subsequent requests",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "api_key": {
                    "description": "Your Captain Data API key",
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "api_key"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session_token": {
                      "description": "Session token to use in Authorization header for subsequent requests",
                      "type": "string"
                    },
                    "expires_in": {
                      "description": "Token expiration time in seconds",
                      "default": 86400,
                      "type": "number"
                    },
                    "_metadata": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "description": "Unique request identifier for tracking",
                          "type": "string"
                        },
                        "executionTime": {
                          "description": "Request execution time in milliseconds",
                          "type": "number"
                        }
                      },
                      "required": [
                        "requestId",
                        "executionTime"
                      ]
                    }
                  },
                  "required": [
                    "session_token",
                    "expires_in",
                    "_metadata"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "description": "Error code for programmatic handling",
                      "type": "string"
                    },
                    "message": {
                      "description": "Human-readable error message",
                      "type": "string"
                    },
                    "requestId": {
                      "description": "Request ID for tracking",
                      "type": "string"
                    },
                    "timestamp": {
                      "format": "date-time",
                      "description": "Error timestamp",
                      "type": "string"
                    },
                    "details": {
                      "description": "Additional error details"
                    },
                    "error_label": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "error_scope": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "input"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "integ"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "param"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "config"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "error_ref": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "status_code": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "params": {
                      "anyOf": [
                        {
                          "type": "object",
                          "additionalProperties": {
                            "type": "string"
                          }
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "description": "Error code for programmatic handling",
                      "type": "string"
                    },
                    "message": {
                      "description": "Human-readable error message",
                      "type": "string"
                    },
                    "requestId": {
                      "description": "Request ID for tracking",
                      "type": "string"
                    },
                    "timestamp": {
                      "format": "date-time",
                      "description": "Error timestamp",
                      "type": "string"
                    },
                    "details": {
                      "description": "Additional error details"
                    },
                    "error_label": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "error_scope": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "input"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "integ"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "param"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "config"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "error_ref": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "status_code": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "params": {
                      "anyOf": [
                        {
                          "type": "object",
                          "additionalProperties": {
                            "type": "string"
                          }
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "description": "Error code for programmatic handling",
                      "type": "string"
                    },
                    "message": {
                      "description": "Human-readable error message",
                      "type": "string"
                    },
                    "requestId": {
                      "description": "Request ID for tracking",
                      "type": "string"
                    },
                    "timestamp": {
                      "format": "date-time",
                      "description": "Error timestamp",
                      "type": "string"
                    },
                    "details": {
                      "description": "Additional error details"
                    },
                    "error_label": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "error_scope": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "input"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "integ"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "param"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "config"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "error_ref": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "status_code": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "params": {
                      "anyOf": [
                        {
                          "type": "object",
                          "additionalProperties": {
                            "type": "string"
                          }
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/introspect": {
      "get": {
        "operationId": "getIntrospect",
        "summary": "List available tools",
        "tags": [
          "Introspection"
        ],
        "description": "Get metadata about available tools (MCP introspection)",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "full"
              ]
            },
            "in": "query",
            "name": "v",
            "required": false,
            "description": "Version parameter. Use \"full\" to get all tools"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tools": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "enum": [
                              "function"
                            ]
                          },
                          "function": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "parameters": {}
                            },
                            "required": [
                              "name",
                              "description",
                              "parameters"
                            ]
                          }
                        },
                        "required": [
                          "type",
                          "function"
                        ]
                      }
                    }
                  },
                  "required": [
                    "tools"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://mcp.captaindata.com",
      "description": "Production server"
    }
  ]
}